+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Add "quick mapping" feature to easily map multiple sources with multiple Solr
+  fields instead of having to create mapping one by one
+
 ## [0.23.0] - 2026-04-28
 
 - Prevent multiple values to be indexed in a non multiValued field